Async updates
DOM updates now happen asynchronously and are batched to be more efficient. When a property of a component is updated it will schedule the DOM update to happen on the next animation frame instead of immediately. Any further property changes in the same task in the event loop will be batched together as one DOM change.

A new global API called Nuro.afterUpdate() has been added to notify the user when all DOM updates have completed.